Shift tests from GemCommandStorageTest to GemCommandTest
