+++ src/test/java/seedu/address/storage/GemCommandStorageTest.java
@@ -137,53 +137,4 @@ public class GemCommandStorageTest {
         ObservableList<Module> actual = storage.getGetModules();
         assertEquals(expected, actual);
     }
-
-    @Test
-    public void moduleExtractor_validTest() throws IOException, DataConversionException {
-        model = new ModelManager();
-        setUpTestModules(SINGLE_MODULE_PATH);
-        String expected = "\n" + MISSING_MODULE_1;
-        StringBuilder temp = gemCommand.moduleExtractor(testModules, model);
-        String actual = "" + temp;
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void getCompiledModules_validTest() {
-        String actual = gemCommand.getCompiledModules();
-        assertNull(actual);
-    }
-
-    @Test
-    public void setCompiledModules_invalidTest() {
-        model = new ModelManager();
-        assertThrows(AssertionError.class, () -> gemCommand.setCompiledModules(storage, model));
-    }
-
-    @Test
-    public void setCompiledModules_validTest() throws IOException, IllegalValueException, DataConversionException {
-        model = new ModelManager();
-
-        String expected = MESSAGE_GEM_SUCCESS + "\n\n" + "Semester 1:" + "\n\n";
-        setUpTestModules(GEH_PATH_1);
-        expected += "Human Cultures\n" + gemCommand.moduleExtractor(testModules, model);
-        setUpTestModules(GET_PATH_1);
-        expected += "\n\nThinking and Expression\n" + gemCommand.moduleExtractor(testModules, model);
-        setUpTestModules(GES_PATH_1);
-        expected += "\n\nSingapore Studies\n" + gemCommand.moduleExtractor(testModules, model);
-        setUpTestModules(GEQ_PATH_1);
-        expected += "\n\nAsking Questions\n" + gemCommand.moduleExtractor(testModules, model);
-        setUpTestModules(GER_PATH_1);
-        expected += "\n\nQuantitative Reasoning\n" + gemCommand.moduleExtractor(testModules, model);
-
-        storage.setGehModules(GEH_SEM1_PATH);
-        storage.setGeqModules(GEQ_PATH);
-        storage.setGerModules(GER_PATH);
-        storage.setGesModules(GES_SEM1_PATH);
-        storage.setGetModules(GET_SEM1_PATH);
-        gemCommand.setCompiledModules(storage, model);
-        String actual = MESSAGE_GEM_SUCCESS + "\n\nSemester 1:\n\n" + gemCommand.getCompiledModules();
-
-        assertEquals(expected, actual);
-    }
 }
